video 10 chapt08

+++ SportStore/SportStore/Models/ViewModels/PagingInfo.cs
@@ -12,7 +12,7 @@ namespace SportStore.Models.ViewModels
         public int CurrentPage { get; set; }
 
         public int TotalPages =>
-            (int)Math.Ceiling((double)TotalItems / ItemsPerPage);
+            (int)Math.Ceiling((decimal)TotalItems / ItemsPerPage);
 
     }
 }
